Factor ChartFieldRangeScaleOptions.tsx out of ChartFieldOption.tsx
